refactor: move helper functions to new utils package
Fixes #120
