+++ ado_asana_sync/sync/sync.py
@@ -14,6 +14,8 @@ from azure.devops.v7_0.work.models import TeamContext
 from azure.devops.v7_0.work_item_tracking.models import WorkItem
 from tinydb import Query
 
+from ado_asana_sync.utils.date import iso8601_utc
+
 from .app import App
 
 # _LOGGER is the logging instance for this file.
@@ -803,21 +805,6 @@ def create_asana_task(
         task.save(a)
     except ApiException as e:
         _LOGGER.error("Exception when calling TasksApi->create_task: %s\n", e)
-
-
-def iso8601_utc(dt: datetime) -> str:
-    """
-    Convert a given datetime object to a string representation in ISO 8601 format with UTC timezone.
-
-    Args:
-        dt (datetime): A datetime object representing a specific date and time.
-
-    Returns:
-        str: A string representing the given datetime object in ISO 8601 format with UTC timezone.
-    """
-    if not dt.tzinfo:
-        dt = dt.replace(tzinfo=timezone.utc)
-    return dt.astimezone(timezone.utc).isoformat()
 
 
 def update_asana_task(a: App, task: TaskItem, tag: TagResponse) -> None:
